+++ brisart_ai/__init__.py
@@ -1,4 +1,4 @@
 """BrisartAI package."""
 
 APP_NAME = "BrisartAI"
-__version__ = "1.0.0-beta.2"
+__version__ = "1.0.0-beta.3"